dynamic allocation of IRDSDataset
Background: [CLIRMatrix](https://www.aclweb.org/anthology/2020.emnlp-main.340/) is [being added to ir_datasets](https://github.com/ssun32/ir_datasets), which contains over 100k sub-datasets. There are so many due to the large number of language combinations. Rather than allocating all of them in the registry (which would be expensive to do every time the library is loaded), we're introducing a "pattern matching" approach that builds them on-the-fly, as they are requested. Thus, not all datasets will appear directly in the registry anymore, which means that they would not be available in pyterrier without a change like this one.