API change: Transformer.transform_iter() returns an iter-dict generator (#481)
(1).  .transform_iter() takes and returns an iterable of dicts generator, rather than a DataFrame
(2) As a consequence of (1), .__call__() takes and returns an iter-dict, takes and returns a iter-dict genetor, OR takes and returns a DataFrame.
(3) Implementers need only implement .transform() OR .transform_iter(). .__call__() redirects based the type of the input.  default implementation is to call .transform_iter() (and vice versa).
(4) Add an easy to remember kwarg for pt.apply to take iter-dicts (iter=True).

Co-authored-by: Sean MacAvaney